+++ ledger/ledger-api-bench-tool/src/main/scala/com/daml/ledger/api/benchtool/metrics/ManagedStreamObserver.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.api.benchtool
+package com.daml.ledger.api.benchtool.metrics
 
 import io.grpc.stub.{ClientCallStreamObserver, ClientResponseObserver}
