+++ ledger/ledger-api-bench-tool/src/main/scala/com/daml/ledger/api/benchtool/services/LedgerIdentityService.scala
@@ -1,11 +1,13 @@
 // Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.api.benchtool
+package com.daml.ledger.api.benchtool.services
 
+import com.daml.ledger.api.v1.ledger_identity_service.{
+  GetLedgerIdentityRequest,
+  LedgerIdentityServiceGrpc,
+}
 import io.grpc.Channel
-import com.daml.ledger.api.v1.ledger_identity_service.LedgerIdentityServiceGrpc
-import com.daml.ledger.api.v1.ledger_identity_service.GetLedgerIdentityRequest
 import org.slf4j.LoggerFactory
 
 final class LedgerIdentityService(channel: Channel) {
